controllers,internal: Move the db package to internal
